backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
pydantic-settings>=2.7.0
python-multipart>=0.0.18
google-genai>=1.0.0
httpx>=0.28.0
pyjwt>=2.10.0
bcrypt>=4.3.0
python-docx>=1.1.0
thefuzz>=0.22.0
python-dotenv>=1.0.0
